Refactor into an App struct with better Ctrl-C handling
